+++ hc/api/views.py
@@ -68,7 +68,7 @@ class Spec(BaseModel):
     methods: Literal["", "POST"] | None = None
     name: str | None = Field(None, max_length=100)
     schedule: str | None = Field(None, max_length=100)
-    slug: str | None = Field(None, pattern="^[a-z0-9-_]*$")
+    slug: str | None = Field(None, max_length=100, pattern="^[a-z0-9-_]*$")
     start_kw: str | None = Field(None, max_length=200)
     subject: str | None = Field(None, max_length=200)
     subject_fail: str | None = Field(None, max_length=200)