+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Remove `pruneflips` management command (now cleaned up automatically)
 - Remove `prunenotifications` management command (now cleaned up automatically)
 - Update settings.py to read SECURE_PROXY_SSL_HEADER from env vars
+- Remove LINE Notify onboarding form (as LINE Notify is shutting down on Apr 1, 2025)
 
 ### Bug Fixes
 - Update sqlite settings to avoid "Database is locked" errors (#1057)